split code further into submodules, fixed imports
