+++ prince/solvers/__init__.py
@@ -1,0 +1,1 @@
+from .propagation import UHECRPropagationResult, UHECRPropagationSolverBDF, UHECRPropagationSolverEULER
